Register Models at definition time
Also added a default namespace of emptystring? At least this way
undefined doesnt get auto casted into 'undefined' because object keys
must be strings.

The nested structure has the advantage of being able to crawl up to find
something, so I think an Object is probably the best if and when a
heavily nested namespaced model is needed. For now though, just one is
supported.