Store the StatementFactories instead of raw sql
This will allow the migrator to convert the managers into SQL using
whatever dialect it requires. This dumbs down the statement factories
such that they only handle converting the user input into a Sequel AST
which the migrator uses to create and run SQL.

The main motivation for this is such that the migrator can have it's own
visitor for the ASTs that will allow it to create a `RuntimeSchema` to
automatically reverse descructive statements like drop_table and
remove_column in the future without the user having to tell the Migrator
what that column was. Just an idea, but saving off the whole
StatementFactory makes a lot of sense right now.