Add RENAME COLUMN to sql
There's a pretty bad unary in the AlterTableManager right now. Columns
can be altered, but when they are renamed they are wrapper in a Unary
node which does not have the alter method. This was a change in order to
differentiate between added and altered columns. Not sure on the best
solution to get rid of the jank